add api and unit tests

+++ database/migrations/2024_06_18_215321_create_order_product_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
-            $table->decimal('quantity', 8, 2);
+            $table->unsignedInteger('quantity');
         });
     }
